+++ custom_components/prix_carburant/const.py
@@ -4,21 +4,25 @@ from typing import Final
 from homeassistant.const import Platform
 
 DOMAIN: Final = "prix_carburant"
-PLATFORMS: Final = [Platform.SENSOR]
+PLATFORMS: Final = [Platform.SENSOR, Platform.BUTTON]
 
 DEFAULT_NAME: Final = "Prix Carburant"
+DEFAULT_MAX_KM: Final = 15
+DEFAULT_SCAN_INTERVAL: Final = 4
 
 ATTR_ADDRESS = "address"
 ATTR_POSTAL_CODE = "postal_code"
+ATTR_BRAND = "brand"
 ATTR_CITY = "city"
-ATTR_FUELS = "fuels"
 ATTR_DISTANCE = "distance"
+ATTR_FUELS = "fuels"
 ATTR_UPDATED_DATE = "updated_date"
 ATTR_DAYS_SINCE_LAST_UPDATE = "days_since_last_update"
 ATTR_PRICE = "price"
-CONF_STATIONS = "stations"
 CONF_MAX_KM = "max_km"
 CONF_FUELS = "fuels"
+CONF_STATIONS = "stations"
+CONF_DISPLAY_ENTITY_PICTURES = "display_entity_pictures"
 
 ATTR_GAZOLE = "Gazole"
 ATTR_SP95 = "SP95"